Merge branch 'bug/rst-forked-gem' into 'master'
Use forked github-markup gem to enable python3 support with omnibus

Changes the version of `github-markup` gem we are using, to be the forked one with the changes to enable python3 support.
The change is backward compatible with users using Python 2, and will only be running with Python 3 in Omnibus (with the help of a custom patch we apply there).

Make sure backward compatibility isn't broken for source installs.

You need to make sure you pickup into stable this merge request as well (for omnibus): https://gitlab.com/gitlab-org/omnibus-gitlab/merge_requests/1015 as it's part of the fix, otherwise package will fail to build.

We had some regressions with .rst rendering that were fixed, but still some issues with encoding persisted (this MR will fix that). See: https://gitlab.com/gitlab-org/gitlab-ce/issues/21696

* gitlab-org/gitlab-ce#21696
* gitlab-org/gitlab-ce#22748
* gitlab-com/support-forum#1097

See merge request !6659

Signed-off-by: Rémy Coutable <remy@rymai.me>