Merge branch '22655-deployments-don-t-always-have-keep-around-refs' into 'master'
Handle case where deployment ref no longer exists

## What does this MR do?

In 8.9, we didn't create keep-around refs for deployments. So it's possible that someone created a deployment (say, for testing), and then deleted the branch and all other references to that commit. That commit could then get GCed, and trying to view MRs on 8.11+ will show a 500. See https://gitlab.com/gitlab-org/gitlab-ce/issues/22655#note_16575020 for more details.

## Why was this MR needed?

If someone created a deployment on 8.9, then deleted all references to the commit for that deployment, we will throw an exception when checking if the deployment includes a commit.

Closes #22655.

See merge request !6855

Signed-off-by: Rémy Coutable <remy@rymai.me>