Merge branch 'fix/fix-public-builds-name-in-pipeline-settings' into 'master'
Fix name of feature that restricts access to builds traces

## What does this MR do?

This MR fixes the name of the feature that makes it possible to restrict build traces. This does not disable access to pipelines. Pipelines are still available so we should tie the name with builds, instead of using work "pipeline" there.

See merge request !6980