Disable warming of the asset cache in Spinach tests under CI
I suspect some combination of Knapsack tests cause no regular Rack tests
to be loaded (i.e. all JavaScript tests), which leads to the error:

ArgumentError: rack-test requires a rack application, but none was given

In CI, we precompile all the assets so there is no need to warm the
asset cache in any case.

Closes #23613