Merge branch 'sh-fix-labels-move-issue' into 'master'
Fix bug where labels would be assigned to issues that were moved

If you attempt to move an issue from one project to another and leave
labels blank, LabelsFinder would assign all labels in the new project
to that issue. The issue is that :title is passed along to the Finder,
but since it appears empty no filtering is done. As a result, all
labels in the group are returned. This fix handles that case.

Closes #23668

See merge request !7065

Signed-off-by: Rémy Coutable <remy@rymai.me>