Move 'Explore Snippets' Spinach feature to Rspec
This commit moves the `snippets/discover.feature` Spinach test to a
Rspec feature, as part of deprecating the Spinach test suite.

The original feature was called 'Discover Snippets', but the UI no
longer reflects this wording. The new Rspec feature is called
'Explore Snippets' to reflect UI/Controller/View naming in use.

- Remove Spinach discover snippets feature and steps
- Add Rspec feature test