Requeue all soft deleted namespaces and projects
The first take on this problem involved a worker requeueing 256 of these
per 4 hours. Than it was suggested that this might not be needed, and
post deployment migrations could be used.

Given we monitor the number of projects which are pending deletion, we
could give this strategy a try and take a look about a few hours later
to see how this fares.