Merge branch '23520-mr-sticky-tabs-overlap-discussion-from-anchor' into 'master'
Account for fixed position MR when scrolling to elements

This MR accounts for the new merge request fixed affix bar when scrolling to an element on the MR page.

The fixed MR tabs bar was not being taken into account when shifting permalink scroll targets so that they are unobscured by navigation elements.

Closes #23520

See merge request !7051

Signed-off-by: Rémy Coutable <remy@rymai.me>