Clarify LDAP troubleshooting ldap_search example and align CE/EE docs [ci skip]
A customer noted an error/lack of clarity in the LDAP documentation
with the `ldap_search` example. Previously, if taken literally, the
customer may have expected the `$` variables to be automatically
replaced or if they paste the exact `user_filter` contents the
parentheses would have been incorrect. Let's just simply the filter
and use exactly what's in the configuration.

Also, this file was way out of wack between CE and EE. I synced them, except
for the banner at the top that links to EE LDAP docs on the EE side.