Rely on Gitlab::Shell public API, not reading internal files it may use
It ensures we rely on `Gitlab::Shell`'s public API and not on its
internal. Since `Gitlab::Shell` is caching the value of its token, the
only way to get the correct token is to call
`Gitlab::Shell.secret_token`, not to read the
`Gitlab.config.gitlab_shell.secret_file` file!

Signed-off-by: Rémy Coutable <remy@rymai.me>