Add active state for outline buttons and update color use
The active state for outline buttons now darkens their background
and border instead of just adding a shadow. Outline buttons now
use a different color for their background and border when they
are in hover, focus or active state. The resting state for outline
buttons now uses the $border-colorname-light colors instead of
$colorname-normal.