Merge branch '24652-fix-timeago-not-rendering-in-repository-view' into 'master'
Fix broken commit metadata loading for large repository trees

## What does this MR do?

Fixes a regression introduced in !7208 in which a call to `ajaxGet()` would fail when loading metadata for large repository trees.

This call produced no errors in the console because it is contained in a sort of `eval` block in jQuery.  This is an ugly, easily broken method for loading data asynchronously whose problems were exemplified by this issue.  I'm opening a separate issue (#24660) to alter this implementation.

## Why was this MR needed?

The failing AJAX call was causing metadata to stop loading, and the `timeago` functionality to stop running due to script failure.

## Screenshots (if relevant)

![Screen_Shot_2016-11-18_at_2.03.13_PM_copy](/uploads/ab0c11d5c405c01d8e45818b3ca96f23/Screen_Shot_2016-11-18_at_2.03.13_PM_copy.png)

## Does this MR meet the acceptance criteria?

- [ ] All builds are passing
- [ ] Branch has no merge conflicts with `master` (if it does - rebase it please)

## What are the relevant issue numbers?

Closes #24652

See merge request !7588