Merge branch 'zj-rake-task-remove-faulty-deployment-refs' into 'master'
Fix faulty deployment refs

In the 8.13 RC cycle, so before the release, there was a time in which
references in git where stored by id instead of iid. This could be fixed
by time, if the iid catches up with the id, it overwrites it. But in the
mean time we have wrong refs in the folder.

This commit fixes that. For all projects we have deployments we'll find
the ones where the ref has a higher number than the iid is now and calls
`#create_ref` on the corresponding deployment.

Fixes gitlab-com/infrastructure#683

cc @ayufan 

One worry: I tested this locally, couldn't find a way to do automate this.

See merge request !7352