Update two_factor_authentication.md
I wanted to practice GFM markdown, so I picked this page more or
less at random. I flattened out the text a little. It's not as
interesting, but will be easier to translate. Use it if it looks
like an improvement. If not, at least I got some practice out of the deal!

Signed-off-by: Rémy Coutable <remy@rymai.me>