Quick fix issuable update when only time spent is updated
Since 9f97fa4d, we don't always set label_ids, meaning that we don't
call `issuable.update` is `params` are empty. The issue is that if you
only send a `/spent 5m` command, it builds an associated `timelogs`
record on the issuable that is only save when we update the issuable.

For now, we check that if we have a `spent_time`, we need to update the
issuable in order to save the associated record. This should be
improved.

Signed-off-by: Rémy Coutable <remy@rymai.me>