Fix failures in default sort order spec
Test cases were not waiting for page load after clicking.
This caused failures in test cases that were executed right after
these incorrect ones, because of requests hitting server after
a test was finished.
We do not have to click links, we can go directly to a correct page
by passing parameters in the URL.