Merge request terms are reversed for GitOperationService
This is seriously confusing but a target branch in merge request,
is actually the source branch for GitOperationService, which is
the base branch. The source branch in a merge request, is the
target branch for GitOperationService, which is where we want to
make commits.

Perhaps we should rename source branch in GitOperationService to
base branch, and target branch to committing branch.