Treat environments matching `production/*` as Production
For Cycle Analytics, the environment `production` is treated as a
special case, and used to update the `first_deployed_to_production_at`
metric.

With the Review Apps changes, slash syntax is used to group related
environments in the list. This is mostly used for `review/<branch>`
environments, but it's possible for people to have multiple production
environments, and someone might then want to group them like
`production/eu`, `production/us`, or even just make it look prettier
with `production/www.example.com`. In any of these cases, the first
deploy to one of these `production/*` environments now is treated as a
Production deploy.