Allow a user to be an "Auditor"
An auditor user is intended to be user with read-only access to all projects and
groups. Access to the admin area and any project settings pages are disallowed

This commit lays the initial groundwork for this concept - adding an `auditor`
column to the `users` table, as well as a few supplements.