Fixed merge request tabs extra margin
Fixed extra margin when the merge request tabs are affixed

This was caused by some weird padding that was on the container & then
not correctly removed.

The fix involved removing the container from inside the tabs holder div
& then adding the widths through CSS depending on if fixed & whether the
parent is limited width or not.

In reality out container classes need to be fixed as it is getting pretty un-managable at the moment this should of been an easy fix, but in reality turned into something much harder.

Closes #26853