Improve performance of triggered chat commands
When the trigger endpoint is called, it has to find the right service
for the given project. However, the old implementation did much more.

For example, it build a list of the missing services on this project.
This whole process took about 750ms _each time_. The current
implementation is expected to perform 10x better, as it only searches in
the current projects services. Given the service has to be configured
anyway, this can be done.