Elasticsearch: allow initial indexing to proceed within Sidekiq
This commit introduces a `gitlab:elastic:index_repositories_async' rake task,
which can be used to enqueue a set of indexing jobs into Sidekiq. These will
be scheduled in the usual manner, in the `sidekiq` queue.

In theory, this allows Sidekiq's concurrency and distributed execution features
to be used to spread and control the load of indexing repositories.